fix: inline measures and repeater core chart
JIRA: F1-236
