+++ libs/sdk-ui-ext/src/internal/components/pluggableVisualizations/repeater/PluggableRepeater.tsx
@@ -6,36 +6,22 @@ import {
     IDimension,
     IInsightDefinition,
     ISettings,
-    IdentifierRef,
     areObjRefsEqual,
-    attributeDisplayFormRef,
     insightBucket,
     insightBuckets,
     insightProperties,
     insightSetBuckets,
-    isMeasure,
-    measureAlias,
-    measureFormat,
-    measureTitle,
-    measureLocalId,
-    modifyInlineMeasure,
-    newInlineMeasure,
-    IMeasure,
-    measureAggregation,
-    measureItem,
-    newBucket,
     IColorMappingItem,
-    isAdhocMeasure,
 } from "@gooddata/sdk-model";
 import {
     RepeaterColumnWidthItem,
     IRepeaterColumnSizing,
-    ChartInlineVisualizationType,
     ColorUtils,
     CoreRepeater,
     IColorMapping,
     constructRepeaterDimensions,
     updateConfigWithSettings,
+    constructRepeaterBuckets,
 } from "@gooddata/sdk-ui-charts";
 import { IExecutionFactory } from "@gooddata/sdk-backend-spi";
 import { BucketNames, IPushData, VisualizationEnvironment } from "@gooddata/sdk-ui";
@@ -122,31 +108,15 @@ export class PluggableRepeater extends AbstractPluggableVisualization {
         const attributeBucket = insightBucket(insight, BucketNames.ATTRIBUTE);
         const viewBucket = insightBucket(insight, BucketNames.VIEW);
         const rowAttribute = attributeBucket.items[0] as IAttribute;
-        const mainRowAttributeRef = attributeDisplayFormRef(rowAttribute) as IdentifierRef;
-        const mainRowAttributeId = mainRowAttributeRef.identifier;
-        const columnBucket = insightBucket(insight, BucketNames.COLUMNS);
+        const columnsBucket = insightBucket(insight, BucketNames.COLUMNS);
         const visualizationProperties = insightProperties(insight);
-        const sanitizedColumnBucketItems = columnBucket.items.map((item) => {
-            if (isMeasure(item)) {
-                const localId = measureLocalId(item);
-                const inlineVisualizationType =
-                    (visualizationProperties?.inlineVisualizations?.[localId]
-                        ?.type as ChartInlineVisualizationType) ?? "metric";
-
-                return transformAdhocMeasureToInline(
-                    item,
-                    inlineVisualizationType === "metric" ? mainRowAttributeId : undefined,
-                );
-            }
-
-            return item;
-        });
-
-        const insightWithSanitizedBuckets = insightSetBuckets(insight, [
-            attributeBucket,
-            { ...columnBucket, items: sanitizedColumnBucketItems },
-            viewBucket ?? newBucket(BucketNames.VIEW),
-        ]);
+        const sanitizedBuckets = constructRepeaterBuckets(
+            rowAttribute,
+            columnsBucket?.items,
+            viewBucket?.items?.[0] as IAttribute,
+            visualizationProperties?.inlineVisualizations,
+        );
+        const insightWithSanitizedBuckets = insightSetBuckets(insight, sanitizedBuckets);
 
         return executionFactory
             .forInsight(insightWithSanitizedBuckets)
@@ -399,36 +369,4 @@ export class PluggableRepeater extends AbstractPluggableVisualization {
             );
         }
     }
-}
-
-export function transformAdhocMeasureToInline(measure: IMeasure, mainRowAttributeId?: string): IMeasure {
-    if (!isAdhocMeasure(measure)) {
-        return measure;
-    }
-
-    const itemRef = measureItem(measure) as IdentifierRef;
-    const aggregation = measureAggregation(measure);
-    let maqlExpression: string;
-
-    const itemIdentifier = `{${itemRef.type === "measure" ? "metric" : itemRef.type}/${itemRef.identifier}}`;
-
-    if (aggregation) {
-        maqlExpression = `SELECT ${aggregation}(${itemIdentifier})`;
-    } else {
-        maqlExpression = `SELECT ${itemIdentifier}`;
-    }
-
-    if (mainRowAttributeId) {
-        maqlExpression += ` BY ALL OTHER EXCEPT {label/${mainRowAttributeId}}`;
-    }
-
-    const inlineMeasure = newInlineMeasure(maqlExpression);
-
-    return modifyInlineMeasure(inlineMeasure, (m) =>
-        m
-            .format(measureFormat(measure))
-            .localId(measureLocalId(measure))
-            .title(measureTitle(measure))
-            .alias(measureAlias(measure)),
-    );
 }
